+++ src/vue/components/Avatar.vue
@@ -1,0 +1,12 @@
+<template>
+  <div class="avatar">
+    <img :src="src.avatar.url" />
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'Avatar',
+  props: ['src']
+}
+</script>